Added buttons to download current parameters as JSON; split InterfaceFunctions to prevent recursive importing; fixed error messages stopping once a severe error was found
