+++ ClientResources/SimulationRunFunctions.py
@@ -21,17 +21,9 @@ from aiohttp import (
 )
 from streamlit_notify import toast  # type: ignore
 
-from ClientResources.InterfaceFunctions import errorChecker, idGet
-from ClientResources.ModelSchema import (
-    Parameters,
-    commandArgument,
-    communityOverride,
-    modelGuideFile,
-    overrideParams,
-    scenarioParameters,
-    simulation,
-    simulationSet,
-)
+from ClientResources.DownloadFunctions import createConfig
+from ClientResources.InterfaceFunctions import errorChecker
+from ClientResources.ParameterFunctions import idGet
 from ClientResources.SharedResources import (
     AnalysisFile,
     ageTimeDict,
@@ -43,12 +35,6 @@ from ClientResources.SharedResources import (
     serverUrl,
     usePresetParams,
 )
-from ParameterTabs.communityParams import communitySchema
-
-# from ParameterTabs.basicParams import basicSchema
-from ParameterTabs.diseaseParams import diseaseSchema
-from ParameterTabs.dynamicParams import dynamicSchema
-from ParameterTabs.vaccinationNPIParams import vaccineSchema
 
 # Logging
 functionLog = logging.getLogger(__name__)
@@ -66,81 +52,6 @@ class invalidSchemaError(Exception):
 
     def __str__(self):
         return f"{self.message} (Full Response: {self.response})"
-
-
-def createConfig(scenarioCount: int):
-    """
-    Function to generate a JSON config file using the selected parameters
-
-    Parameters:
-        scenarioCount (int): The number of scenarios to define in the config.
-    """
-    # Set up schema objects
-    scenarioParams = [Parameters() for _ in range(scenarioCount)]
-
-    # Populate parameters with session_state values
-    useVaccines = False
-    useAdvanced = session.get("showAdvanced", False)
-    for id, scenario in enumerate(scenarioParams):
-        diseaseSchema(scenario, id, useAdvanced)
-        communitySchema(scenario, id, useAdvanced)
-        useVaccines = vaccineSchema(scenario, id, useAdvanced) or useVaccines
-        if useAdvanced:
-            dynamicSchema(scenario, id)
-
-    # Use middle joint to control options
-    # TODO: Account for more conditionals
-    middleJoint = "-dashboard"
-    if useVaccines:
-        middleJoint += "+vaccines"
-
-    # Create config object
-    return modelGuideFile(
-        name="Flusim Dashboard Simulation",
-        description=str(session.sessionID),
-        output_folder="./results/",
-        middle_joint=middleJoint,
-        community_used=[session.get("community", "newcastle")],
-        # Community overrides are global parameters e.g. number of runs
-        community_overrides=[
-            communityOverride(
-                name=session.get("community", "newcastle"),
-                parameters=Parameters(
-                    Command_Argument=commandArgument(
-                        n_runs=session.get("runCount", 24),
-                        n_cycles=session.get("cycleCount", 360) * 2,
-                    ),
-                    Scenario_Parameter=scenarioParameters(
-                        start_day_of_week=(
-                            "Sunday",
-                            "Monday",
-                            "Tuesday",
-                            "Wednesday",
-                            "Thursday",
-                            "Friday",
-                            "Saturday",
-                        ).index(session.get("startDay", "Monday"))
-                    ),
-                ),
-            )
-        ],
-        # Shared overrides are baseline parameters
-        shared_overrides=overrideParams(parameters=scenarioParams[0]),
-        simulation_sets=[
-            simulationSet(
-                name="Dashboard Simulation Set",
-                version=session.sessionID,
-                simulations=[simulation(name="Baseline")]
-                + [
-                    simulation(
-                        name=session[f"scenarioName{i}"],
-                        override_setting=overrideParams(parameters=scenarioParams[i]),
-                    )
-                    for i in range(1, scenarioCount)
-                ],
-            )
-        ],
-    )
 
 
 @st.dialog("Run Simulation Experiment", width="large", icon=":material/motion_play:")
@@ -183,11 +94,14 @@ community data to simulate each of the following {scenarioCount + 1} scenarios:
     # TODO: Hide scenario errors that are copies of baseline errors
     severeErrorsFound = False
     for id in range(scenarioCount + 1):
-        severeErrorsFound = severeErrorsFound or errorChecker(
-            id,
-            f"""Errors in {
-                session[f'scenarioName{id}'] if id > 0 else 'Baseline'
-            }""",
+        severeErrorsFound = (
+            errorChecker(
+                id,
+                f"""
+                    Errors in {session[f'scenarioName{id}'] if id > 0 else 'Baseline'}
+                """,
+            )
+            or severeErrorsFound
         )
     if severeErrorsFound:
         st.error(
