+++ DashboardPages/runSimulations.py
@@ -8,12 +8,8 @@ import logging
 import streamlit as st
 
 # from streamlit_push_notifications import send_push, send_alert
-from ClientResources.InterfaceFunctions import (
-    dayCount,
-    loadKey,
-    saveKey,
-    timeScaleChange,
-)
+from ClientResources.InterfaceFunctions import dayCount
+from ClientResources.ParameterFunctions import loadKey, saveKey, timeScaleChange
 from ClientResources.SharedResources import communityPopulation
 from ClientResources.SimulationRunFunctions import runSimulationButton
 
@@ -48,19 +44,19 @@ community = st.selectbox(
     on_change=saveKey,
     args=["community"],
     help="""
-        The Australian city whose community data will be used as the
-        basis for the population and demographic distribution in the
-        simulation. Note that the data used for these communities comes
-        from 2011.
+The Australian city whose community data will be used as the
+basis for the population and demographic distribution in the
+simulation. Note that the data used for these communities comes
+from 2011.
 
-        ##### Options:
-        - Newcastle: A metropolitan area in New South Wales, Australia.
-        It has a population of 272407, the second-largest in the state,
-        and has a demographic distribution that more closely matches
-        that of Australia as a whole compared to Cairns.
-        - Cairns: A major city in Queensland, Australia. It has a
-        population of 140402 (as of 2011 when this data was collected)
-        and has a higher Indigenous population compared to Newcastle.
+##### Options:
+- Newcastle: A metropolitan area in New South Wales, Australia.
+It has a population of 272407, the second-largest in the state,
+and has a demographic distribution that more closely matches
+that of Australia as a whole compared to Cairns.
+- Cairns: A major city in Queensland, Australia. It has a
+population of 140402 (as of 2011 when this data was collected)
+and has a higher Indigenous population compared to Newcastle.
     """,
 )
 
@@ -74,12 +70,12 @@ st.select_slider(
     key="_cycleCount",
     on_change=timeScaleChange,
     help="""
-        The length of the time period that will be simulated, measured in days.
+The length of the time period that will be simulated, measured in days.
 
-        Note that if you lower this value, other time-based parameters
-        may have their values altered. For instance, if you go from 360
-        days to 120, a NPI set to end on Day 180 will be changed to end
-        on Day 120 instead.
+Note that if you lower this value, other time-based parameters
+may have their values altered. For instance, if you go from 360
+days to 120, a NPI set to end on Day 180 will be changed to end
+on Day 120 instead.
     """,
 )
 
@@ -93,10 +89,10 @@ st.slider(
     on_change=saveKey,
     args=["runCount"],
     help="""
-        How many times each scenario will be simulated. The results
-        of each individual simulation will be averaged together to
-        get the final results; higher values lead to longer
-        simulations but more accurate results.
+How many times each scenario will be simulated. The results
+of each individual simulation will be averaged together to
+get the final results; higher values lead to longer
+simulations but more accurate results.
     """,
 )
 
@@ -109,8 +105,7 @@ st.select_slider(
     on_change=saveKey,
     args=["startDay"],
     help="""
-        The day of the week that the first day of the
-        simulation will be.
+The day of the week that the first day of the simulation will be.
     """,
 )
 
@@ -131,17 +126,17 @@ st.button(
     icon=("spinner" if session.simulationInProgress else ":material/motion_play:"),
     help=(
         """
-        Send a request to the *Flusim* model server to run the model
-        with the specified parameters. Once the request has been made,
-        you will be unable to run the model again until it completes,
-        so make sure you have configured your parameters to appropriate
-        values before clicking.
-    """
+Send a request to the *Flusim* model server to run the model
+with the specified parameters. Once the request has been made,
+you will be unable to run the model again until it completes,
+so make sure you have configured your parameters to appropriate
+values before clicking.
+        """
         if not session.simulationInProgress
         else """
-        A simulation is already running; please wait for it to conclude
-        before running another one.
-    """
+A simulation is already running; please wait for it to conclude
+before running another one.
+        """
     ),
 )
