+++ DashboardPages/scenarioParameters.py
@@ -7,12 +7,9 @@ import logging
 
 import streamlit as st
 
-from ClientResources.InterfaceFunctions import (
-    containerSave,
-    errorChecker,
-    idGet,
-    loadKey,
-)
+from ClientResources.DownloadFunctions import parameterDownload
+from ClientResources.InterfaceFunctions import errorChecker
+from ClientResources.ParameterFunctions import containerSave, idGet, loadKey
 from ClientResources.SharedResources import maxScenarios
 from ParameterTabs.communityParams import buildCommunityTab
 from ParameterTabs.diseaseParams import buildDiseaseTab
@@ -177,23 +174,25 @@ showAdvanced = st.toggle(
     args=["showAdvanced"],
     kwargs={"containers": containersToOpen},
     help="""
-        Toggle whether to display parameters that control more fine-grain
-        aspects of the simulation environment, such as age-specific NPI
-        compliance or dynamic parameter updates.
+Toggle whether to display parameters that control more fine-grain
+aspects of the simulation environment, such as age-specific NPI
+compliance or dynamic parameter updates.
     """,
 )
+
+# TODO: Buttons to upload simulation parameters
+parameterDownload()
 
 for id in range(1, scenarioCount + 1):
     # TODO: Consider changing expanders to popovers or tabs
     # to avoid the nested expander issue
     containerScenarioName = session.get(f"scenarioName{id}", f"Scenario #{id}")
-    scenarioExpander = st.expander(
+    with st.expander(
         f"#{id}: {containerScenarioName}",
         key=f"scenarioContainer{id}",
         on_change="rerun",
-    )
-    if scenarioExpander.open:
-        with scenarioExpander:
+    ) as scenarioExpander:
+        if scenarioExpander.open:
             st.header(f"Scenario #{id}")
             # Scenario name
             loadKey("scenarioName", id, f"Scenario #{id}")
@@ -207,8 +206,8 @@ for id in range(1, scenarioCount + 1):
                 args=["scenarioName", id, [f"scenarioContainer{id}"]],  # type: ignore
                 placeholder="Enter a name for this scenario",
                 help="""
-                    The name to give to this scenario, which will display
-                    in tables and graphs generated by the dashboard.
+The name to give to this scenario, which will display
+in tables and graphs generated by the dashboard.
                 """,
             )
             # Parameters for this scenario
@@ -270,9 +269,8 @@ for id in range(1, scenarioCount + 1):
                 on_click=deleteScenario,
                 args=[id],
                 help="""
-                    Remove this scenario from the simulation set, thus
-                    ensuring that it is not ran when you run the
-                    simulation.
+Remove this scenario from the simulation set, thus ensuring that it is not
+ran when you run the simulation.
                 """,
             )
 
@@ -286,15 +284,13 @@ st.button(
     disabled=not scenarioCount < maxScenarios,
     help=(
         """
-        Add another scenario to the simulation, where you can configure
-        different parameter values to use instead of the baseline
-        values.
-    """
+Add another scenario to the simulation, where you can configure
+different parameter values to use instead of the baseline values.
+        """
         if scenarioCount < maxScenarios
         else f"""
-        To keep the number of scenarios manageable, no more than {maxScenarios}
-        scenarios plus the baseline may be added to the simulation set
-        at once.
+To keep the number of scenarios manageable, no more than {maxScenarios}
+scenarios plus the baseline may be added to the simulation set at once.
     """
     ),
 )
